Cleaned up DirectionsGenerator.java A LOT fixed a typo in one of the comments of Node.java
In DirectionsGenerator.java, I added comments to every method, as well as wrote helper methods for the giant method.
Though the helper methods did not reduce the size of the string generator, they greatly increased the readability.
In addition to those, I also added additional improvements to the text directions.